+++ client/src/components/App.scss
@@ -1,4 +1,4 @@
-@import "variables.scss";
+@import "../styles/variables.scss";
 
 //TODO: Some scss needs to be moved down into the appropriate scss files
