+++ client/src/components/Channel/ChannelList.scss
@@ -1,0 +1,1 @@
+@import "../App.scss";
