+++ client/src/components/Channel/ChannelListItem.scss
@@ -1,4 +1,4 @@
-@import "./variables.scss";
+@import "../../styles/variables.scss";
 
 .channel-title {
   margin: 1rem;
@@ -22,7 +22,6 @@
   margin-top: 3rem;
   align-self: center;
   color: white;
-  
 }
 
 .channel-icon {
