--- client/src/components/ComponentTemplate.jsx
@@ -1,6 +1,0 @@
-const X = (props) => {
-  const {} = props;
-  return;
-};
-
-export default X;
