+++ client/src/components/Login/Header.jsx
@@ -1,6 +1,6 @@
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import SettingsIcon from "@mui/icons-material/Settings";
-import "../styles/Header.scss";
+import "./Header.scss";
 // TODO: Header needs actual interactive icons rather than the placeholder text here, register/login/logout needs to be implemented (Probably create a new component that get rendered by the Header and handles that)
 
 const Header = (props) => {
