+++ client/src/components/Login/Header.scss
@@ -1,4 +1,4 @@
-@import "variables.scss";
+@import "../../styles/variables.scss";
 
 .container {
   height: 100%;
