+++ client/src/components/Login/Login.scss
@@ -1,4 +1,5 @@
-@import "./variables.scss";
+@import "../../styles/variables.scss";
+
 .login-page {
   display: flex;
   flex-direction: column;
