+++ client/src/components/Message/ChatInput.scss
@@ -1,4 +1,4 @@
-@import "./variables.scss";
+@import "../../styles/variables.scss";
 
 .input-form {
   display: flex;
