+++ client/src/components/Message/MessageListItem.scss
@@ -1,4 +1,4 @@
-@import "./variables.scss";
+@import "../../styles/variables.scss";
 
 .message-item {
   display: flex;
