+++ client/src/components/Room/RoomListItem.jsx
@@ -1,6 +1,6 @@
 import FiberManualRecordIcon from "@mui/icons-material/FiberManualRecord";
 import classNames from "classnames";
-import "../styles/RoomListItem.scss";
+import "./RoomListItem.scss";
 
 // TODO: RoomListItem needs new message notifications implemented, consider incrementing a message count as new messages come in and when the user is in the room currently don't increment the counter or reset it or something
