+++ client/src/components/Room/RoomListItem.scss
@@ -1,4 +1,4 @@
-@import "./variables.scss";
+@import "../../styles/variables.scss";
 
 .room-container {
   display: flex;
