+++ client/src/components/User/FriendListItem.scss
