--- client/src/styles/ChannelList.scss
@@ -1,1 +1,0 @@
-@import "./App.scss";
